result-service/requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
python-jose
python-dotenv